requirements.txt
flask
google-generativeai
pillow
python-dotenv